http: Move web server to core and use flags.
This way we can independently control the debugger and disc sharing, since
debugger could have security impact.
