+++ UI/RemoteISOScreen.h
@@ -25,8 +25,6 @@
 #include "UI/MiscScreens.h"
 #include "UI/MainScreen.h"
 
-bool StartRemoteISOSharing();
-
 class RemoteISOScreen : public UIScreenWithBackground {
 public:
 	RemoteISOScreen();
